feat: implement feedback [Finishes #151550754]

+++ server/controllers/userController.js
@@ -1,6 +1,7 @@
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
 import env from 'dotenv';
+import util from 'util';
 import User from '../models/user';
 import UsersGroup from '../models/usersgroup';
 
@@ -16,7 +17,7 @@ export default class UserController {
  * @param {obj} req
  * @param {obj} res
  * @param {obj} next
- * @return {JSON} Returns success or failure message with the data
+ * @return {obj} Returns success or failure message with the user token
  *
  */
   static signup(req, res) {
@@ -32,7 +33,8 @@ export default class UserController {
           User.create({
             name, username, email, phone, password: hash
           }).then((user) => {
-            const payload = { username: user.username,
+            const payload = {
+              username: user.username,
               userId: user.id,
               email: user.email,
               name: user.name
@@ -41,20 +43,14 @@ export default class UserController {
               expiresIn: 60 * 60 * 24
             });
             res.status(200).json({
-              data: {
-                id: user.id,
-                name: user.name,
-                username: user.username,
-                email: user.email,
-                phone: user.phone
-              },
               message: 'Account created',
               token
             });
           }).catch((err) => {
             if (err) {
               res.status(409).json({
-                message: 'Record exists already' });
+                message: 'Record exists already'
+              });
             }
           });
         });
@@ -67,7 +63,7 @@ export default class UserController {
  * @param {obj} req
  * @param {obj} res
  * @param {obj} next
- * @return {obj} Return success or failure message
+ * @return {obj} Return user token for successful login or error response
  */
   static signin(req, res) {
     const username = req.body.username.toLowerCase(),
@@ -75,7 +71,8 @@ export default class UserController {
     User.findOne({ where: { username } }).then((user) => {
       if (user && user.dataValues.username === username) {
         const check = bcrypt.compareSync(password, user.dataValues.password);
-        const payload = { username: user.dataValues.username,
+        const payload = {
+          username: user.dataValues.username,
           userId: user.dataValues.id,
           email: user.dataValues.email,
           name: user.dataValues.name
@@ -85,22 +82,25 @@ export default class UserController {
             expiresIn: 60 * 60 * 24
           });
           res.status(200).json({
-            data: {
-              userId: user.dataValues.id,
-              username: user.dataValues.username,
-              email: user.dataValues.email,
-
+            user: {
+              id: user.id,
+              name: user.name,
+              username: user.username,
+              email: user.email,
+              phone: user.phone
             },
             message: 'Logged In',
             token
           });
         } else {
           res.status(400).json({
-            message: 'Invalid Password' });
+            message: 'Invalid Password'
+          });
         }
       } else {
         res.status(404).json({
-          message: 'User not found' });
+          message: 'User not found'
+        });
       }
     });
   }
@@ -114,28 +114,34 @@ export default class UserController {
   static addUser(req, res) {
     const groupId = req.params.groupId,
       username = req.body.username;
-    User.findOne({ where: { username } }).then((user) => {
-      if (!user) {
-        res.json({ message: 'Username does not exits' });
-      } else {
-        return UsersGroup.sync({ force: false }).then(() => {
-          UsersGroup.findOrCreate({ where: { userId: user.id, groupId } })
-          .spread((usergg, created) => {
-            if (created) {
-              res.status(200).json({
-                message: 'User successfully added' });
-            } else {
-              res.status(409).json({ status: 'failed',
-                message: 'User already exist in this group' });
-            }
+    const isGroupId = Number.isInteger(parseInt(groupId, 10));
+    if (isGroupId) {
+      User.findOne({ where: { username } }).then((user) => {
+        if (!user) {
+          res.json({ message: 'Username does not exits' });
+        } else {
+          return UsersGroup.sync({ force: false }).then(() => {
+            UsersGroup.findOrCreate({ where: { userId: user.id, groupId } })
+              .spread((usergg, created) => {
+                if (created) {
+                  res.status(200).json({
+                    message: 'User successfully added'
+                  });
+                } else {
+                  res.status(409).json({
+                    status: 'failed',
+                    message: 'User already exist in this group'
+                  });
+                }
+              });
           });
-        });
+        }
       }
+      );
     }
-    );
   }
   /**
-   * @return {array} Returns array of objects
+   * @return {array} Returns arrays of all users in a group
    * @param {obj} req
    * @param {obj} res
    */
@@ -164,32 +170,40 @@ export default class UserController {
     }
   }
 
-/**
- * @return {obj} Returns object containing pageCount and arrays of users
- * @param {*} req
- * @param {*} res
- */
-  static getUsers(req, res) {
-    const search = req.query.search;
-    const limit = req.query.limit;
-    const offset = req.query.offset || 0;
+  /**
+   * @return {obj} Returns object containing pageCount and arrays of users and searchMetadata
+   * @param {*} req
+   * @param {*} res
+   */
+  static searchUsers(req, res) {
+    const PER_PAGE = 5;
+    const offset = req.query.offset ? parseInt(req.query.offset, 10) : 0,
+      nextOffset = offset + PER_PAGE,
+      previousOffset = (offset - PER_PAGE < 1) ? 0 : offset - PER_PAGE;
+    const search = req.query.search.toLowerCase();
     if (search) {
       User.findAndCountAll({
+        attributes: ['id', 'name', 'username'],
         where: {
           username: {
             $like: `%${search}%`
           }
-        },
-        offset,
-        limit
+        }
       })
-.then((result) => {
-  const pageCount = Math.ceil(result.count / 5),
-    users = result.rows;
-  res.status(200).json({ pageCount, users });
-});
+        .then((result) => {
+          const users = result.rows;
+          const paginatedUsers = users.slice(offset, offset + PER_PAGE);
+          const searchMetaData = {
+            limit: PER_PAGE,
+            next: util.format('?limit=%s&offset=%s', PER_PAGE, nextOffset),
+            offset: req.query.offset,
+            previous: util.format('?limit=%s&offset=%s', PER_PAGE, previousOffset),
+            total_count: result.count
+          };
+          res.status(200).json({ searchMetaData, paginatedUsers });
+        });
     } else {
-      res.status(200).json({ pageCount: 0, users: [] });
+      res.status(200).json({ searchMetaData: {}, paginatedUsers: [] });
     }
   }
 
@@ -204,7 +218,8 @@ export default class UserController {
       if (!user) {
         User.sync({ force: false }).then(() => {
           User.create({ name, username, email }).then((userDetail) => {
-            const payload = { username: userDetail.username,
+            const payload = {
+              username: userDetail.username,
               userId: userDetail.id,
               email: userDetail.email,
               name: userDetail.name
@@ -225,7 +240,8 @@ export default class UserController {
           });
         });
       } else {
-        const payload = { username: user.dataValues.username,
+        const payload = {
+          username: user.dataValues.username,
           userId: user.dataValues.id,
           email: user.dataValues.email,
           name: user.dataValues.name
